Changing to double precision.  - Remaining parts in single precision are the mesh coordinates (since they are only available in SP from the Solver)  - Binaries are also written in single precision  - NetCDF files from the Solver contain everything in SP.

+++ fft_type.f90
@@ -241,7 +241,7 @@ end function
 
 !-----------------------------------------------------------------------------------------
 function taperandzeropad(array, ntimes)
-  real(kind=sp), intent(in)     :: array(:,:)
+  real(kind=dp), intent(in)     :: array(:,:)
   integer,       intent(in)     :: ntimes
   real(kind=dp)                 :: taperandzeropad(ntimes, size(array,2))
   real(kind=dp), allocatable    :: win(:,:)
